be/requirements.txt
django-cors-headers==4.3.0
django==4.2.6
langchain==0.0.281
python-dotenv==1.0.0
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0